UX: make loading visible
